+++ src/main/java/me/phyrextsai/hyperledger/crowdfunding/utils/CampaignUtils.java
@@ -2,6 +2,7 @@ package me.phyrextsai.hyperledger.crowdfunding.utils;
 
 import me.phyrextsai.hyperledger.crowdfunding.data.Campaign;
 import me.phyrextsai.hyperledger.crowdfunding.data.Contribute;
+import me.phyrextsai.hyperledger.crowdfunding.helper.CampaignHelper;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.hyperledger.java.shim.ChaincodeStub;
@@ -39,33 +40,14 @@ public class CampaignUtils {
     }
 
     public String init() {
-        Campaign campaign = new Campaign();
-        boolean campaignCreateSuccess = campaign.create(stub);
+        CampaignHelper helper = new CampaignHelper();
+        boolean campaignCreateSuccess = helper.create(stub);
         if (campaignCreateSuccess) {
-            System.out.println(String.format("Create table [%s]", Campaign.CAMPAIGN));
+            System.out.println(String.format("Create table [%s]", CampaignHelper.CAMPAIGN));
         } else {
-            System.out.println(String.format("Table [%s] exist", Campaign.CAMPAIGN));
+            System.out.println(String.format("Table [%s] exist", CampaignHelper.CAMPAIGN));
         }
         return null;
-    }
-
-    public String create(String[] args) {
-        try {
-            String campaignId = UUID.randomUUID().toString();
-            String address = args[0];
-            String info = args[1];
-            Integer fundingAmount = Integer.parseInt(args[2]);
-            Campaign campaign = new Campaign(campaignId, address, info, fundingAmount);
-
-            if (campaign.insert(stub, campaign)) {
-                log.info("Insert record, CampaignId : " + campaignId);
-                return "{\"Data\":\"Create Campaign success, uuid : " + campaignId + ".\"}";
-            } else {
-                return "{\"Error\":\"Create Campaign failed.\"}";
-            }
-        } catch (NumberFormatException e) {
-            return "{\"Error\":\"Expecting integer value for asset holding\"}";
-        }
     }
 
     public String owner(String campaignId) {
@@ -86,7 +68,7 @@ public class CampaignUtils {
         List<TableProto.Column> key = new ArrayList<>();
         key.add(queryCol);
         try {
-            TableProto.Row tableRow = stub.getRow(Campaign.CAMPAIGN,key);
+            TableProto.Row tableRow = stub.getRow(CampaignHelper.CAMPAIGN,key);
             if (tableRow.getSerializedSize() > 0) {
                 // TODO: better use getColumnList to check
                 for(TableProto.Column col : tableRow.getColumnsList()){
@@ -94,7 +76,7 @@ public class CampaignUtils {
                 }
                 return tableRow.getColumns(column).getString();
             } else {
-                return String.format("Can not found %s record!", Campaign.CAMPAIGN);
+                return String.format("Can not found %s record!", CampaignHelper.CAMPAIGN);
             }
         } catch (Exception invalidProtocolBufferException) {
             invalidProtocolBufferException.printStackTrace();
